web demo/BE/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==9.0.0
numpy==1.21.0 